add readwrite config file
